Split main into client, srv; move GameState to model
