+++ mpmp/src/client/ClientMain.java
@@ -1,14 +1,8 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package main;
+package client;
 
 import java.io.IOException;
 import java.net.UnknownHostException;
 
-import controller.Controller;
 import model.Card;
 
 public class ClientMain {
